requirements.txt
discord.py
aiosqlite